requirements.txt
aws-cdk-lib>=2.202.0
constructs>=10.4.2